# Conflicts:
#	src/main/java/tech/dat250project/controller/PollController.java

+++ src/main/java/tech/dat250project/controller/PollController.java
@@ -21,7 +21,9 @@ import tech.dat250project.repository.PersonRepository;
 import tech.dat250project.repository.PollRepository;
 import tech.dat250project.security.UserDetailsImpl;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @RestController
@@ -139,19 +141,25 @@ public class PollController {
                     content = {@Content(mediaType = "application/json",
                             schema = @Schema(implementation = Poll.class))})
     })
-    Poll update(@RequestBody Poll newPoll, @PathVariable Long id) {
+    ResponseEntity update(@RequestBody Poll newPoll, @PathVariable Long id) {
+        UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Person person = personRepository.findByUsername(userDetails.getUsername()).orElse(null);
         return pollRepository.findById(id)
                 .map(poll -> {
-                    poll.setQuestion(newPoll.getQuestion());
-                    poll.setStatus(newPoll.getStatus());
-                    if (poll.getOpened() != newPoll.getOpened()) {
-                        dweetPoster.publish(newPoll);
-                        sender.send(poll);
+                    if(Objects.equals(person.getId(), poll.getAuthor().getId())) {
+                        if (poll.getOpened() != newPoll.getOpened()) {
+                            dweetPoster.publish(newPoll);
+                            sender.send(poll);
+                        }
+                        poll.setQuestion(newPoll.getQuestion());
+                        poll.setStatus(newPoll.getStatus());
+                        poll.setOpened(newPoll.getOpened());
+                        return ResponseEntity.ok(pollRepository.save(poll));
+                    } else {
+                        return ResponseEntity.badRequest().body(new Message("You are not author of this poll!"));
                     }
-                    poll.setOpened(newPoll.getOpened());
-                    return pollRepository.save(poll);
                 })
-                .orElseGet(() -> pollRepository.save(newPoll));
+                .orElseGet(() -> ResponseEntity.ok(pollRepository.save(new Poll(newPoll.getQuestion(), newPoll.getOpened(), newPoll.getStatus(), person))));
     }
 
     @Operation(summary = "Deletes a poll given its id")
@@ -162,38 +170,72 @@ public class PollController {
             @ApiResponse(responseCode = "404", description = "Poll not found",
                     content = @Content)
     })
-    void delete(@PathVariable Long id) {
+    ResponseEntity delete(@PathVariable Long id) {
+        UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Person person = personRepository.findByUsername(userDetails.getUsername()).orElse(null);
+        Poll poll = pollRepository.findById(id).orElse(null);
+        if(poll == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Message("Poll not found!"));
+        if(!Objects.equals(person.getId(), poll.getAuthor().getId())) return ResponseEntity.badRequest().body(new Message("You are not author of this poll!"));
         pollRepository.deleteById(id);
+        return ResponseEntity.ok(new Message("Successfully deleted."));
     }
 
-    @Operation(summary = "Attaches one or more devices to a poll given its id")
-    @PostMapping("/poll/{id}/device")
+    @Operation(summary = "Gets one or more devices to a poll given its id")
+    @GetMapping("/poll/{code}/device")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Device(s) attached successfully",
+            @ApiResponse(responseCode = "200", description = "Device(s) of poll",
                     content = {@Content(mediaType = "application/json",
                             schema = @Schema(implementation = ResponseEntity.class))}),
             @ApiResponse(responseCode = "404", description = "Poll not found",
                     content = {@Content(mediaType = "application/json",
                             schema = @Schema(implementation = ResponseEntity.class))})
     })
-    ResponseEntity<HttpStatus> assignDevices(@RequestBody List<Integer> ids, @PathVariable Long id) {
-        Poll poll = pollRepository.findById(id).orElse(null);
+    ResponseEntity getDevices(@PathVariable String code) {
+        Poll poll = pollRepository.findByCode(code).orElse(null);
+        List<Device> devices = new ArrayList<>();
         if(poll != null) {
-            Device device;
-            for (Integer deviceId : ids) {
-                device = deviceRepository.findById(deviceId.longValue()).orElse(null);
-                if(device != null) {
+            List<DevicePoll> devicePolls = devicePollRepository.findByPollId(poll.getId());
+            for(DevicePoll devicePoll:devicePolls) {
+                devices.add(devicePoll.getDevice());
+            }
+            return ResponseEntity.ok(devices);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Message("Poll not found!"));
+    }
+
+    @Operation(summary = "Attaches one device to a poll given its id")
+    @PostMapping("/poll/{code}/device")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Device attached successfully",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ResponseEntity.class))}),
+            @ApiResponse(responseCode = "404", description = "Poll or device not found",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ResponseEntity.class))})
+    })
+    ResponseEntity assignDevices(@RequestBody Long id, @PathVariable String code) {
+        UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Person person = personRepository.findByUsername(userDetails.getUsername()).orElse(null);
+        Poll poll = pollRepository.findByCode(code).orElse(null);
+        if(poll != null) {
+            if(Objects.equals(poll.getAuthor().getId(), person.getId())) {
+                Device device;
+                device = deviceRepository.findById(id).orElse(null);
+                if (device != null) {
                     DevicePoll devicePoll = new DevicePoll(device, poll);
                     devicePollRepository.save(devicePoll);
+                    return ResponseEntity.ok(device);
                 }
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Message("Device not found!"));
+            } else {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Message("You are not author!"));
             }
-            return new ResponseEntity<>(HttpStatus.OK);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Message("Poll not found!"));
     }
 
     @Operation(summary = "Deletes the link between a poll and a device given its id")
-    @DeleteMapping("/poll/{id}/device/{deviceId}")
+    @DeleteMapping("/poll/{code}/device/{deviceId}")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Device detached from poll successfully",
                     content = {@Content(mediaType = "application/json",
@@ -202,12 +244,22 @@ public class PollController {
                     content = {@Content(mediaType = "application/json",
                             schema = @Schema(implementation = ResponseEntity.class))})
     })
-    ResponseEntity<HttpStatus> deleteDevice(@PathVariable Long deviceId, @PathVariable Long id) {
-        DevicePoll devicePoll = devicePollRepository.findByDeviceIdAndPollId(deviceId, id);
-        if(devicePoll != null) {
-            devicePollRepository.delete(devicePoll);
-            return new ResponseEntity<>(HttpStatus.OK);
+    ResponseEntity deleteDevice(@PathVariable Long deviceId, @PathVariable String code) {
+        UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Person person = personRepository.findByUsername(userDetails.getUsername()).orElse(null);
+        Poll poll = pollRepository.findByCode(code).orElse(null);
+        if(poll != null) {
+            if(Objects.equals(poll.getAuthor().getId(), person.getId())) {
+                DevicePoll devicePoll = devicePollRepository.findByDeviceIdAndPollId(deviceId, poll.getId());
+                if (devicePoll != null) {
+                    devicePollRepository.delete(devicePoll);
+                    return ResponseEntity.ok(new Message("Device unlinked."));
+                }
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Message("Device is not linked to this poll!"));
+            } else {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Message("You are not author!"));
+            }
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new Message("Poll not found!"));
     }
 }
